+++ packages/components/CHANGELOG.md
@@ -78,7 +78,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - ⌛Input Date
 - ✅Input Email
 - ⌛Input Number
-- ⌛Input Password
+- ✅Input Password
 - ✅Input Search
 - ✅Input Tel
 - ✅Input Text